first implementation of settickets sys call (task 3)
no testing yet, as task 4 is not completed. may need to add code
inside sysproc.c settickets() to modify values of global_tickets,
global_stride, and global_pass, as these values are not changed
after modifying the procs ticket value. however, these seem to be
recomputed whenever they are used, so this might not be necessary, idk